Refactor game model to reflect changes to game tbale schema
