+++ server/src/models/index.js
@@ -1,5 +1,5 @@
 // include all of your models here using CommonJS requires
 const User = require("./User.js")
-const OneUserGame = require("./OneUserGame.js")
+const Game = require("./Game.js")
 
 module.exports = {User};
